fixed #4
added static function that search given name in a given list of string
added maven support to project
