+++ .gitignore
@@ -1,1 +1,3 @@
 /.metadata/
+
+\.idea/
